app: extract actions into separate functions
