+++ internal/app/export.go
@@ -3,9 +3,13 @@ package app
 import (
 	"context"
 	"errors"
+	"fmt"
+	"os"
 
 	"github.com/rusq/slackdump/internal/export"
 )
+
+const defDirMode = 0700
 
 // Export performs the full export of slack workspace in slack export compatible
 // format.
@@ -14,6 +18,9 @@ func (app *App) Export(ctx context.Context, dir string) error {
 		return errors.New("export directory not specified")
 	}
 
+	if err := os.MkdirAll(dir, defDirMode); err != nil {
+		return fmt.Errorf("Export: failed to create the export directory %q: %w", dir, err)
+	}
 	export := export.New(dir, app.sd)
 	return export.Run(ctx)
 }
