+++ internal/export/conversation_test.go
@@ -12,19 +12,23 @@ import (
 )
 
 func TestConversation_ByDate(t *testing.T) {
+	// TODO
 	var exp Export
-	c := fixtures.Load[slackdump.Conversation](fixtures.TestConversationJSON)
-	convDt, err := exp.byDate(&c, fixtures.Load[slackdump.Users](fixtures.UsersJSON))
+	conversations := fixtures.Load[slackdump.Conversation](fixtures.TestConversationJSON)
+	convDt, err := exp.byDate(&conversations, fixtures.Load[slackdump.Users](fixtures.UsersJSON))
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	data, err := json.Marshal(convDt)
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	if err := os.WriteFile("x.json", data, 0644); err != nil {
 		t.Fatal(err)
 	}
+
 	assert.Equal(t, map[string][]slack.Conversation{}, convDt)
 }
